Save variables as URL parameters for restoration
Save variables as URL parameters for restoration:
https://git.kiwifarms.net/wolfree/wolfree.sh/src/commit/48e0272abc61cbe1eda5e7c67079d68eb44847ee/wolfree.sh
All files in this directory are output from wolfree.sh.
Running wolfree.sh should automatically generate all the files in this directory.
The source code for wolfree.sh is available on this Gitea server:
https://git.kiwifarms.net/wolfree
Alternatively, use the Tor Browser to access the hidden Gitea service:
http://it7otdanqu7ktntxzm427cba6i53w6wlanlh23v5i3siqmos47pzhvyd.onion/wolfree