Doc: automatically generate a table of contents.
Documentation: Automatically generate a table of contents.
Click on the triple bar to expand the table of contents.
All files in this commit are output from wolfree.sh.
Running wolfree.sh should automatically generate all the files in this commit.
The source code for wolfree.sh is available on the Gitea server:
https://git.kiwifarms.net/wolfree/wolfree.sh/src/commit/628c6d02d8849da94d7f092799cf05b2cee7d3d3/wolfree.sh