List more Fediverse accounts and Gitea servers.
List more Fediverse accounts and Gitea servers.

The following archives are the source code for this commit.
https://archive.ph/X5NJ9
https://web.archive.org/web/20220906171807/https://git.disroot.org/wolfree/wolfree-dockerfile/src/commit/57a65d87e31f141f328345818976ce9468681aa6/Dockerfile

All the files in this commit are created automatically using Docker and Dockerfile.

Docker can read Dockerfile to produce all the files in this commit.

You can get Docker from the following website.
https://docs.docker.com/

You can get Dockerfile from the following Gitea servers.
https://git.disroot.org/wolfree
https://git.kiwifarms.net/wolfree
http://it7otdanqu7ktntxzm427cba6i53w6wlanlh23v5i3siqmos47pzhvyd.onion/wolfree

Cloudflare has blocked Kiwifarms. Visitors to any of the Kiwifarms sites that use any of Cloudflare's services will see a Cloudflare block page and a link to the following post.
https://blog.cloudflare.com/kiwifarms-blocked/

Before the commit, only the following two Gitea servers were listed in the documentation.
https://git.kiwifarms.net/wolfree
http://it7otdanqu7ktntxzm427cba6i53w6wlanlh23v5i3siqmos47pzhvyd.onion/wolfree

After the commit, the following Gitea server is also listed in the documentation.
https://git.disroot.org/wolfree

When git.kiwifarms.net is not accessible, you can use Disroot's Gitea server.

The reason for adding more Fediverse accounts is similar.

Before the commit, only the following two Fediverse accounts were listed in the documentation.
https://poa.st/@wolfree
https://kiwifarms.cc/wolfree

After the commit, the following Fediverse account is also listed in the documentation.
https://freespeechextremist.com/wolfree